+++ src/CacheControl/CacheControl.php
@@ -2,30 +2,28 @@
 
 namespace Nuwave\Lighthouse\CacheControl;
 
-final class CacheControl
+class CacheControl
 {
     /**
      * List of maxAges.
      *
-     * @var array<int>
+     * @var array<int, int>
      */
-    private $maxAgeList = [];
+    protected $maxAgeList = [];
 
     /**
      * List of scopes.
      *
-     * @var array<string>
+     * @var array<int, string>
      */
-    private $scopeList = [];
+    protected $scopeList = [];
 
-    public function addToMaxAgeList(int $maxAge): void
-    {
+    public function addToMaxAgeList(int $maxAge): void {
         $this->maxAgeList[] = $maxAge;
     }
 
-    public function addToScopeList(string $scope): void
-    {
-        $this->scopeList[] = $scope;
+    public function addToScopeList(string $scope): void {
+        $this->scopeList[] = strtolower($scope);
     }
 
     /**
@@ -33,62 +31,27 @@ final class CacheControl
      *
      * @return array<string, bool|int>
      */
-    public function makeHeaderOptions(): array
-    {
+    public function makeHeaderOptions(): array {
         // Early return if @cacheControl not used.
         if (empty($this->scopeList) && empty($this->maxAgeList)) {
-            return $this->appDefaultHeader();
+            $headerOptions['no_cache'] = true;
+            $headerOptions['private'] = true;
+
+            return $headerOptions;
         }
 
-        $maxAge = $this->calculateMaxAge();
-        // Set max-age value.
+        $maxAge = min($this->maxAgeList);
         if (0 === $maxAge) {
             $headerOptions['no_cache'] = true;
         } else {
             $headerOptions['max_age'] = $maxAge;
         }
 
-        // Set scope value.
-        if ('public' === $this->calculateScope()) {
-            $headerOptions['public'] = true;
-        } else {
+        if (in_array('private', $this->scopeList)) {
             $headerOptions['private'] = true;
+        } else {
+            $headerOptions['public'] = true;
         }
-
-        return $headerOptions;
-    }
-
-    /**
-     * Return the minimum maxAge from the list.
-     */
-    private function calculateMaxAge(): int
-    {
-        assert(! empty($this->maxAgeList));
-
-        return min($this->maxAgeList);
-    }
-
-    /**
-     * Return scope as public unless there is a private scope.
-     */
-    private function calculateScope(): string
-    {
-        if (in_array('PRIVATE', $this->scopeList)) {
-            return 'private';
-        }
-
-        return 'public';
-    }
-
-    /**
-     * App default behavior is no-cache, private.
-     *
-     * @return array<string, bool>
-     */
-    private function appDefaultHeader(): array
-    {
-        $headerOptions['no_cache'] = true;
-        $headerOptions['private'] = true;
 
         return $headerOptions;
     }