+++ src/Support/Http/Controllers/GraphQLController.php
@@ -5,7 +5,6 @@ namespace Nuwave\Lighthouse\Support\Http\Controllers;
 use Illuminate\Contracts\Events\Dispatcher as EventsDispatcher;
 use Illuminate\Http\Request;
 use Laragraph\Utils\RequestParser;
-use Nuwave\Lighthouse\CacheControl\CacheControl;
 use Nuwave\Lighthouse\Events\EndRequest;
 use Nuwave\Lighthouse\Events\StartRequest;
 use Nuwave\Lighthouse\GraphQL;
@@ -21,8 +20,7 @@ class GraphQLController
         EventsDispatcher $eventsDispatcher,
         RequestParser $requestParser,
         CreatesResponse $createsResponse,
-        CreatesContext $createsContext,
-        CacheControl $cacheControl
+        CreatesContext $createsContext
     ): Response {
         $eventsDispatcher->dispatch(
             new StartRequest($request)
@@ -34,13 +32,10 @@ class GraphQLController
         $result = $graphQL->executeOperationOrOperations($operationOrOperations, $context);
 
         $response = $createsResponse->createResponse($result);
-        $response = $response->setCache($cacheControl->makeHeaderOptions());
 
         $eventsDispatcher->dispatch(
             new EndRequest($response)
         );
-
-        app()->forgetInstance(CacheControl::class);
 
         return $response;
     }