+++ CHANGELOG.md
@@ -11,7 +11,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ### Added
 
-- Add `@cacheControl` directive
+- Add `@cacheControl` directive https://github.com/nuwave/lighthouse/pull/2132
 
 ## v5.48.1